Added SpriteComponent; fixed some little issues

+++ src/common/GameObject.cpp
@@ -20,7 +20,7 @@ GameObject::~GameObject()
 
 shared_ptr<IComponent> GameObject::getComponent(unsigned id)
 {
-    for(int i = 0; i < _componentArray.size(); ++i)
+    for(size_t i = 0; i < _componentArray.size(); ++i)
     {
         if(id == _componentArray[i].type)
             return ComponentFabric::instance()->getComponent(_componentArray[i]);
@@ -45,7 +45,7 @@ void GameObject::addComponent(unsigned type)
 
 bool GameObject::hasComponent(unsigned type) const
 {
-    for(int i = 0; i < _componentArray.size(); ++i)
+    for(size_t i = 0; i < _componentArray.size(); ++i)
     {
         if(type == _componentArray[i].type)
             return true;
